+++ src/Helpers/Entity/Discount.php
@@ -1,5 +1,4 @@
 <?php
-
 declare(strict_types=1);
 
 namespace MarcelStrahl\PriceCalculator\Helpers\Entity;
@@ -7,29 +6,18 @@ namespace MarcelStrahl\PriceCalculator\Helpers\Entity;
 use MarcelStrahl\PriceCalculator\Helpers\Types\DiscountInterface;
 
 /**
- * Class Discount
  * @author Marcel Strahl <info@marcel-strahl.de>
- * @package MarcelStrahl\PriceCalculator\Helpers\Entity
  */
 class Discount implements DiscountInterface
 {
-    /**
-     * @var int
-     */
-    private int $percent;
+    private float $percent;
 
-    /**
-     * @param int $percent
-     */
-    public function __construct(int $percent)
+    public function __construct(float $percent)
     {
         $this->percent = $percent;
     }
 
-    /**
-     * {@inheritdoc}
-     */
-    public function getDiscount(): int
+    public function getDiscount(): float
     {
         return $this->percent;
     }
